增加了虚拟乐器开启关闭（渐隐效果）功能 增加了移除乐器功能 限制了ventruecmd只能从ventrue获取 修改了包络控制decay阶段的控制函数 增加了移除乐器对应设备通道 增加新增乐器功能 增加替换原始乐器功能

+++ Ventrue/src/core/Synth/Channel.h
@@ -15,7 +15,7 @@ namespace ventrue
 		~Channel();
 
 		//获取通道号
-		inline int GetChannelNum()
+		inline int64_t GetChannelNum()
 		{
 			return channelNum;
 		}
@@ -109,7 +109,7 @@ namespace ventrue
 
 	private:
 		//通道号
-		int channelNum = 0;
+		int64_t channelNum = 0;
 		//所在轨道
 		Track* track = nullptr;
